refactor: FileRepositoryImp.kt deleteFile function

+++ app/src/main/java/com/saboon/project_2511sch/data/repository/FileRepositoryImp.kt
@@ -10,7 +10,9 @@ import com.saboon.project_2511sch.util.Resource
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
+import java.io.IOException
 import javax.inject.Inject
+import java.io.File as JavaFile
 
 class FileRepositoryImp @Inject constructor(
     private val fileDao: FileDao,
@@ -26,6 +28,10 @@ class FileRepositoryImp @Inject constructor(
 
     override suspend fun deleteFile(file: File): Resource<File> {
         try {
+            val fileToDelete = JavaFile(file.filePath)
+            if (fileToDelete.exists() && !fileToDelete.delete()){
+                return Resource.Error("Failed to delete physical file at ${file.filePath}")
+            }
             fileDao.delete(file.toEntity())
             return Resource.Success(file)
         }catch (e: Exception){
